test(T-481): stop exec_module-ing a sibling test file at import time
test_loop_session_limit.py exec_module'd test_weekend_model_ladder.py for
LADDER/LOOPS/_run, so renaming the ladder file was a collection ERROR and
two module objects of every ladder test class lived in one process.

RED (simulated rename to test_ladder.py):
  FileNotFoundError: .../scripts/tests/test_weekend_model_ladder.py
  Interrupted: 1 error during collection

GREEN: shared harness (LOOPS, LADDER, quota lines, _clone, _stub_bin, _run)
moved to non-test scripts/tests/_loop_harness.py; both files import it
(conftest adds scripts/tests to sys.path). Same simulated rename now
collects 20 tests. test_loop_session_limit.py: 20 passed.
test_weekend_model_ladder.py: 69 passed.
